+++ libuavcan/src/transport/uc_transfer_buffer.cpp
@@ -216,7 +216,7 @@ int DynamicTransferBufferManagerEntry::write(unsigned offset, const uint8_t* dat
     }
 
     const int actually_written = len - left_to_write;
-    max_write_pos_ = max(offset + actually_written, unsigned(max_write_pos_));
+    max_write_pos_ = max(uint16_t(offset + actually_written), uint16_t(max_write_pos_));
     return actually_written;
 }
 
@@ -260,7 +260,7 @@ int StaticTransferBufferImpl::write(unsigned offset, const uint8_t* data, unsign
     }
     UAVCAN_ASSERT((offset + len) <= size_);
     (void)copy(data, data + len, data_ + offset);
-    max_write_pos_ = max(offset + len, unsigned(max_write_pos_));
+    max_write_pos_ = max(uint16_t(offset + len), uint16_t(max_write_pos_));
     return len;
 }
 
@@ -268,7 +268,7 @@ void StaticTransferBufferImpl::reset()
 {
     max_write_pos_ = 0;
 #if UAVCAN_DEBUG
-    fill(data_, data_ + size_, 0);
+    fill(data_, data_ + size_, uint8_t(0));
 #endif
 }
 
@@ -306,7 +306,7 @@ bool StaticTransferBufferManagerEntryImpl::migrateFrom(const TransferBufferManag
         TransferBufferManagerEntry::reset();
         return false;
     }
-    buf_.setMaxWritePos(res);
+    buf_.setMaxWritePos(uint16_t(res));
     if (res < int(buf_.getSize()))
     {
         return true;
@@ -329,7 +329,7 @@ StaticTransferBufferManagerEntryImpl* TransferBufferManagerImpl::findFirstStatic
 {
     for (unsigned i = 0; true; i++)
     {
-        StaticTransferBufferManagerEntryImpl* const sb = getStaticByIndex(i);
+        StaticTransferBufferManagerEntryImpl* const sb = getStaticByIndex(uint16_t(i));
         if (sb == NULL)
         {
             break;
@@ -492,7 +492,7 @@ unsigned TransferBufferManagerImpl::getNumStaticBuffers() const
     unsigned res = 0;
     for (unsigned i = 0; true; i++)
     {
-        StaticTransferBufferManagerEntryImpl* const sb = getStaticByIndex(i);
+        StaticTransferBufferManagerEntryImpl* const sb = getStaticByIndex(uint16_t(i));
         if (sb == NULL)
         {
             break;
